requirement.txt
fastapi 
uvicorn
pymongo
pydantic
passlib
python-dotenv
passlib[bcrypt]
python-jose[cryptography]